app/requirements-app.txt
fastapi==0.115.0
uvicorn==0.29.0
pydantic==2.11.0
pydantic-settings==2.12.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
requests==2.32.3
tzlocal==5.3.1
scikit-learn==1.7.0
pandas==2.3.0
joblib==1.5.1
openai==2.14.0
email-validator==2.3.0
asyncpg==0.29.0
sqlalchemy==2.0.41
sqlalchemy[asyncio]
redis==7.1.0
python-multipart==0.0.9